+++ Sources/Swift/Protocol/SentryRedactOptions.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 @objc
-protocol SentryRedactOptions {
+public protocol SentryRedactOptions {
     var maskAllText: Bool { get }
     var maskAllImages: Bool { get }
     var maskedViewClasses: [AnyClass] { get }
@@ -10,8 +10,8 @@ protocol SentryRedactOptions {
 
 @objcMembers
 final class SentryRedactDefaultOptions: NSObject, SentryRedactOptions {
-    var maskAllText: Bool = true
-    var maskAllImages: Bool = true
-    var maskedViewClasses: [AnyClass] = []
-    var unmaskedViewClasses: [AnyClass] = []
+    public var maskAllText: Bool = true
+    public var maskAllImages: Bool = true
+    public var maskedViewClasses: [AnyClass] = []
+    public var unmaskedViewClasses: [AnyClass] = []
 }