+++ Sources/Swift/Tools/SentryViewScreenshotProvider.swift
@@ -3,7 +3,7 @@
 import Foundation
 import UIKit
 
-typealias ScreenshotCallback = (UIImage) -> Void
+public typealias ScreenshotCallback = (UIImage) -> Void
 
 @objc
 protocol SentryViewScreenshotProvider: NSObjectProtocol {